Fix skill rounding error

+++ src/creatures/players/vocations/vocation.cpp
@@ -95,7 +95,7 @@ bool Vocations::loadFromXml()
 		}
 
 		if ((attr = vocationNode.attribute("manamultiplier"))) {
-			voc.manaMultiplier = pugi::cast<float>(attr.value());
+			voc.manaMultiplier = pugi::cast<double>(attr.value());
 		}
 
 		if ((attr = vocationNode.attribute("attackspeed"))) {
@@ -124,7 +124,7 @@ bool Vocations::loadFromXml()
 				if (skillIdAttribute) {
 					uint16_t skill_id = pugi::cast<uint16_t>(skillIdAttribute.value());
 					if (skill_id <= SKILL_LAST) {
-						voc.skillMultipliers[skill_id] = pugi::cast<float>(childNode.attribute("multiplier").value());
+						voc.skillMultipliers[skill_id] = pugi::cast<double>(childNode.attribute("multiplier").value());
 					} else {
 						SPDLOG_WARN("[Vocations::loadFromXml] - "
                                     "No valid skill id: {} for vocation: {}",
@@ -208,7 +208,7 @@ uint64_t Vocation::getReqSkillTries(uint8_t skill, uint32_t level)
 	uint64_t& tries = skillCache[level];
 
 	if (tries == 0) {
-		tries = std::max<uint64_t>(1, static_cast<uint64_t>(skillBase[skill] * fast_pow(static_cast<double>(skillMultipliers[skill]), level - 1)));
+		tries = std::max<uint64_t>(1, static_cast<uint64_t>(skillBase[skill] * fast_pow(skillMultipliers[skill], level - 1)));
 	}
 	return tries;
 }
@@ -225,7 +225,7 @@ uint64_t Vocation::getReqMana(uint32_t magLevel)
 
 	uint64_t& tries = manaCache[magLevel];
 	if (tries == 0) {
-		tries = std::max<uint64_t>(1, static_cast<uint64_t>(1600 * fast_pow(static_cast<double>(manaMultiplier), magLevel - 1)));
+		tries = std::max<uint64_t>(1, static_cast<uint64_t>(1600 * fast_pow(manaMultiplier, magLevel - 1)));
 	}
 	return tries;
 }
